feat: added automatically adding newly generated trained models files feature
